ice_demo/requirements.txt
redis>=2.10.5
zeroc-ice==3.6.1
requests
